# FileStream rewrite: Caching the ValueTaskSource in AsyncWindowsFileStreamStrategy