--- a/src/libraries/System.Private.CoreLib/src/System/IO/Strategies/BufferedFileStreamStrategy.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/IO/Strategies/BufferedFileStreamStrategy.cs
@@ -604,6 +604,7 @@ namespace System.IO.Strategies
             }
             else
             {
+                Debug.Assert(_writePos == _bufferSize);
                 FlushWrite();
             }