… src/libraries/System.Private.CoreLib/src/System/IO/Strategies/FileStreamStrategy.cs