Fixes https://github.com/dotnet/runtime/issues/50972
Fixes #25074

When `AsyncWindowsFileStreamStrategy` is wrapped by a `BufferedFileStreamStrategy`, we need to make sure the `ValueTaskSource` instance is cached to reduce the number of allocations when calling `ReadAsync` or `WriteAsync` multiple times in a row.

This PR is a continuation of https://github.com/dotnet/runtime/pull/50802, in which we switched from using `TaskCompletionSource` to `IValueTaskSource`.

Changes:
- Moved the `PreAllocatedOverlapped` instance inside `ValueTaskSource`, so the latter becomes its owner. This was done because we are only supposed to have an instance of a `PreAllocatedOverlapped` if the `ValueTaskSource` was created from `OnBufferedAllocated`, which is a method called only by `BufferedFileStreamStrategy` right before writing or reading.
- Removed `MemoryValueTaskSource` and moved the cases handled by it to `ValueTaskSource`.
- Created a method that refreshes the value of the `NativeOverlapped*`. This is done every time we call `ReadAsync`/`WriteAsync`, to make sure we are pinning the memory passed by the user.