@adamsitnik 2021-04-16
Initial benchmark results (base is #50802):

<details>

|                 Method |          Toolchain |  fileSize | userBufferSize |      options |          Mean | Ratio |     Gen 0 | Allocated |
|----------------------- |------------------- |---------- |--------------- |------------- |--------------:|------:|----------:|----------:|
|              ReadAsync | \cache\corerun.exe |      1024 |           1024 | Asynchronous |      83.99 us |  1.02 |    0.3360 |      5 KB |
|              ReadAsync |  \base\corerun.exe |      1024 |           1024 | Asynchronous |      81.98 us |  1.00 |    0.6545 |      5 KB |
|                        |                    |           |                |              |               |       |           |           |
|             WriteAsync | \cache\corerun.exe |      1024 |           1024 | Asynchronous |     487.70 us |  1.02 |         - |      5 KB |
|             WriteAsync |  \true\corerun.exe |      1024 |           1024 | Asynchronous |     476.89 us |  1.00 |         - |      5 KB |
|                        |                    |           |                |              |               |       |           |           |
|              ReadAsync | \cache\corerun.exe |   1048576 |            512 | Asynchronous |   2,649.16 us |  1.09 |   10.4167 |    127 KB |
|              ReadAsync |  \true\corerun.exe |   1048576 |            512 | Asynchronous |   2,421.94 us |  1.00 |    8.9286 |     83 KB |
|                        |                    |           |                |              |               |       |           |           |
|             WriteAsync | \cache\corerun.exe |   1048576 |            512 | Asynchronous |   4,527.33 us |  1.11 |   15.6250 |    119 KB |
|             WriteAsync |  \true\corerun.exe |   1048576 |            512 | Asynchronous |   4,100.58 us |  1.00 |         - |     75 KB |
|                        |                    |           |                |              |               |       |           |           |
|              ReadAsync | \cache\corerun.exe |   1048576 |           4096 | Asynchronous |   2,321.45 us |  1.01 |    8.9286 |     71 KB |
|              ReadAsync |  \true\corerun.exe |   1048576 |           4096 | Asynchronous |   2,308.26 us |  1.00 |    8.9286 |     69 KB |
|                        |                    |           |                |              |               |       |           |           |
|             WriteAsync | \cache\corerun.exe |   1048576 |           4096 | Asynchronous |   4,191.59 us |  1.03 |         - |     97 KB |
|             WriteAsync |  \true\corerun.exe |   1048576 |           4096 | Asynchronous |   4,077.05 us |  1.00 |         - |     74 KB |
|                        |                    |           |                |              |               |       |           |           |
|  ReadAsync_NoBuffering | \cache\corerun.exe |   1048576 |          16384 | Asynchronous |     730.67 us |  1.00 |         - |     18 KB |
|  ReadAsync_NoBuffering |  \true\corerun.exe |   1048576 |          16384 | Asynchronous |     732.58 us |  1.00 |         - |     17 KB |
|                        |                    |           |                |              |               |       |           |           |
| WriteAsync_NoBuffering | \cache\corerun.exe |   1048576 |          16384 | Asynchronous |   2,747.11 us |  0.99 |         - |     18 KB |
| WriteAsync_NoBuffering |  \true\corerun.exe |   1048576 |          16384 | Asynchronous |   2,787.66 us |  1.00 |         - |     17 KB |
|                        |                    |           |                |              |               |       |           |           |
|              ReadAsync | \cache\corerun.exe | 104857600 |           4096 | Asynchronous | 250,998.78 us |  1.01 |         - |  7,001 KB |
|              ReadAsync |  \true\corerun.exe | 104857600 |           4096 | Asynchronous | 249,280.08 us |  1.00 |         - |  6,801 KB |
|                        |                    |           |                |              |               |       |           |           |
|             WriteAsync | \cache\corerun.exe | 104857600 |           4096 | Asynchronous | 382,734.73 us |  1.04 | 1000.0000 |  9,205 KB |
|             WriteAsync |  \true\corerun.exe | 104857600 |           4096 | Asynchronous | 368,403.91 us |  1.00 |         - |  6,905 KB |
|                        |                    |           |                |              |               |       |           |           |
|  ReadAsync_NoBuffering | \cache\corerun.exe | 104857600 |          16384 | Asynchronous |  80,072.69 us |  0.97 |         - |  1,750 KB |
|  ReadAsync_NoBuffering |  \true\corerun.exe | 104857600 |          16384 | Asynchronous |  82,376.92 us |  1.00 |         - |  1,700 KB |
|                        |                    |           |                |              |               |       |           |           |
| WriteAsync_NoBuffering | \cache\corerun.exe | 104857600 |          16384 | Asynchronous | 120,385.98 us |  0.98 |         - |  1,751 KB |
| WriteAsync_NoBuffering |  \true\corerun.exe | 104857600 |          16384 | Asynchronous | 123,215.33 us |  1.00 |         - |  1,701 KB |

</details>

it looks like we are allocating less when buffering is disabled, but more than before when it's enabled