@stephentoub 2021-04-16
The interlocked operation that was trying to enable the instance to be reused was backwards... nothing was being pooled.  The numbers will need to be re-run after that's fixed ;-)